backend/functions/requirements.txt
firebase_admin==7.1.0
firebase_functions==0.4.3
langchain==0.3.27
langchain_core==0.3.74
langchain_openai==0.3.31
langchain_pinecone==0.2.11
openai==1.101.0
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
